Allow more concurrent connections
We consistently run out of connections on this server, yet it's plausible the server can handle more connections and this hard limit we've configured is unnecessarily low.
See https://fly.io/docs/apps/concurrency/ for more info on concurrency. There is no documented ceiling for fly.io that I can find.

We'll have to monitor the instances in production to see if this new upper limit becomes a problem or not. Load testing is too much work for something we just provide as a convenience to the community.